Reduced code duplication with some fixtures moved to conftest.py
